% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_metadata.R
\name{add_BIDS_metadata}
\alias{add_BIDS_metadata}
\title{This function adds all required BIDS metadata: 'CHANGES', 'README', 'dataset_description.json', 'participants.json', 'participants.tsv' and per functional sequence a file that contains the RepetitionTime.}
\usage{
add_BIDS_metadata()
}
\description{
This function adds all required BIDS metadata: 'CHANGES', 'README', 'dataset_description.json', 'participants.json', 'participants.tsv' and per functional sequence a file that contains the RepetitionTime.
}
