% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_metadata.R, R/functions_copy2bids.R
\name{add_participants_tsv}
\alias{add_participants_tsv}
\title{This function creates the required BIDS files 'participants.tsv' and 'participants.json'. The '.tsv' contains the actual data per subject from the 'json_metadata.tsv'. The '.json' contains a description of the extracted variables. The following variables are extracted: subject, session, group, PatientBirthDate, AcquisitionDateTime, Patientsex, PatientWeight.}
\usage{
add_participants_tsv(
  tsv_path = paste0(path_output_converter, "/json_metadata.tsv")
)

add_participants_tsv(
  tsv_path = paste0(path_output_converter, "/json_metadata.tsv")
)
}
\arguments{
\item{tsv_path}{the path to the 'json_metadata.tsv'}
}
\value{



}
\description{
This function creates the required BIDS files 'participants.tsv' and 'participants.json'. The '.tsv' contains the actual data per subject from the 'json_metadata.tsv'. The '.json' contains a description of the extracted variables. The following variables are extracted: subject, session, group, PatientBirthDate, AcquisitionDateTime, Patientsex, PatientWeight.

This function creates the required BIDS files 'participants.tsv' and 'participants.json'. The '.tsv' contains the actual data per subject from the 'json_metadata.tsv'. The '.json' contains a description of the extracted variables. The following variables are extracted: subject, session, group, PatientBirthDate, AcquisitionDateTime, Patientsex, PatientWeight.
}
