% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytable.R
\name{check_BIDS_plausibility}
\alias{check_BIDS_plausibility}
\title{Checks the plausibility of entered BIDS sequences based on regular expressions.}
\usage{
check_BIDS_plausibility(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Checks the plausibility of entered BIDS sequences based on regular expressions.
}
