% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{check_folder_order}
\alias{check_folder_order}
\title{Checks the folder order, reads the directory, extracts session and subject folder.}
\usage{
check_folder_order()
}
\description{
Checks the folder order, reads the directory, extracts session and subject folder.
}
