% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input.R
\name{check_folder_order_df}
\alias{check_folder_order_df}
\title{Checks the folder order}
\usage{
check_folder_order_df(df, input_order = user_input_order)
}
\arguments{
\item{input_order}{}
}
\value{

}
\description{
Checks the folder order
}
