% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sequences.R
\name{check_sequence_map}
\alias{check_sequence_map}
\title{Checks sequence map}
\usage{
check_sequence_map(sequence_map_file = "sequence_map")
}
\arguments{
\item{sequence_map_file}{}
}
\description{
Checks sequence map
}
