% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{cleaning_session_ids}
\alias{cleaning_session_ids}
\title{Wrapper function for the editing of session-IDS}
\usage{
cleaning_session_ids()
}
\description{
Wrapper function for the editing of session-IDS
}
