% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input.R
\name{cleaning_subject_ids}
\alias{cleaning_subject_ids}
\title{Function to clean the input subject-ID's from redundant prefixes, suffices or strings.}
\usage{
cleaning_subject_ids()
}
\value{

}
\description{
Function to clean the input subject-ID's from redundant prefixes, suffices or strings.
}
