% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_setup.R
\name{convert_to_BIDS}
\alias{convert_to_BIDS}
\title{The core function, which converts the DICOM to NII, extracts the JSON information, opens the sequence mapper and converts the data to BIDS.}
\usage{
convert_to_BIDS(user_settings_file = settings_file, sequence_table = "off")
}
\arguments{
\item{user_settings_file}{The path to the user settings file from the "create_user_settungs()" function.}

\item{sequence_table}{"on" or "off" - turns the sequence mapper shiny app on or off.}
}
\value{

}
\description{
The core function, which converts the DICOM to NII, extracts the JSON information, opens the sequence mapper and converts the data to BIDS.
}
