% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_copy2bids.R
\name{copy2BIDS}
\alias{copy2BIDS}
\title{Prepares and copies files to BIDS}
\usage{
copy2BIDS(sequence_map = "sequence_map", sequence_file = "json_metadata_anon")
}
\arguments{
\item{tsv_file}{}
}
\value{

}
\description{
Prepares and copies files to BIDS
}
