% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{copy_files}
\alias{copy_files}
\title{Copies files, if they don't exist in target}
\usage{
copy_files(from, to, string)
}
\arguments{
\item{from}{A list: File source}

\item{to}{A list: File destination}

\item{string}{Which step?}
}
\description{
Copies files, if they don't exist in target
}
\examples{
\dontrun{copy_files(path_input, path_output, "Copy files to output.")}
}
