% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{create_BIDS_regex}
\alias{create_BIDS_regex}
\title{Creates the BIDS regular expression}
\usage{
create_BIDS_regex()
}
\value{
one regular expression containing the BIDS nomenclature
}
\description{
Creates the BIDS regular expression
}
