% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{create_dashboard}
\alias{create_dashboard}
\title{Creates the internal dashboard. This one contains sensitive participant information like age, sex and ID.}
\usage{
create_dashboard(
  rmd_file = system.file("rmd", "bids_dashboard.Rmd", package = "bidsconvertr"),
  converter_path = path_output_converter
)
}
\arguments{
\item{rmd_file}{The rmd file to render}

\item{converter_path}{The path to the converter, which contains the bids files}
}
\value{

}
\description{
Creates the internal dashboard. This one contains sensitive participant information like age, sex and ID.
}
