% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_settings.R
\name{create_environment_variables}
\alias{create_environment_variables}
\title{Sets up the environment, creates needed variables variables}
\usage{
create_environment_variables(input_path = path_output)
}
\arguments{
\item{input_path}{The path for the BIDSconvertR output}
}
\description{
Sets up the environment, creates needed variables variables
}
