% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input.R
\name{create_subject_session_df}
\alias{create_subject_session_df}
\title{Extracts the input data folder structure}
\usage{
create_subject_session_df(directory = user_input_dir)
}
\arguments{
\item{directory}{}
}
\description{
Extracts the input data folder structure
}
