% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{create_taskname_metadata}
\alias{create_taskname_metadata}
\title{Extracts the sequence id and RepetitionTime of functional sequences for the required BIDS files.}
\usage{
create_taskname_metadata(
  tsv_path = paste0(path_output_converter, "/identifying_information/json_metadata.tsv"),
  tsv_map = paste0(path_output_converter, "/sequence_map.tsv")
)
}
\arguments{
\item{tsv_path}{the path to the 'json_metadata.tsv'}

\item{tsv_map}{the path to the 'sequence_map.tsv'}
}
\description{
Extracts the sequence id and RepetitionTime of functional sequences for the required BIDS files.
}
