% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{create_user_settings}
\alias{create_user_settings}
\title{Creates a user setting template file at working directory or custom folder. This file is required for all processing.}
\usage{
create_user_settings(
  folder = shinyDirectoryInput::choose.dir(caption =
    "Select the folder, where your 'user_settings.R' file should be saved.")
)
}
\arguments{
\item{folder}{Set the folder to save your file.}
}
\value{
The path to the folder
}
\description{
Creates a user setting template file at working directory or custom folder. This file is required for all processing.
}
