% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dicom_wrapper.R
\name{dcm2nii_converter}
\alias{dcm2nii_converter}
\title{dcm2niix system calls using a list from dcm2nii_wrapper}
\usage{
dcm2nii_converter(dcm2niix_string, input_folder, output_folder)
}
\arguments{
\item{output_folder}{list of output folders (one for each subject and session)}

\item{list}{from dcm2nii_wrapper}
}
\description{
dcm2niix system calls using a list from dcm2nii_wrapper
}
\examples{
\dontrun{dcm2nii_converter("dcm2niix -o nii/session-id/participant-id/ -ba y -f \%d -z y root_folder/session_id/participant_id/")}
}
