% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{dcm2nii_converter_anon}
\alias{dcm2nii_converter_anon}
\title{Converts the DICOM to anonymized NII and JSON files.}
\usage{
dcm2nii_converter_anon()
}
\description{
Converts the DICOM to anonymized NII and JSON files.
}
