% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dicom_wrapper.R
\name{dcm2nii_converter_json}
\alias{dcm2nii_converter_json}
\title{Converts the DICOM to only JSON files, that contain all sensitive information stored in the DICOM headers.}
\usage{
dcm2nii_converter_json()
}
\value{

}
\description{
Converts the DICOM to only JSON files, that contain all sensitive information stored in the DICOM headers.
}
