% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{dcm2nii_wrapper}
\alias{dcm2nii_wrapper}
\title{Creates the dcm2niix system commands for the conversion}
\usage{
dcm2nii_wrapper(
  input_folder,
  output_folder,
  scanner_type,
  dcm2niix_local_path = normalizePath(dcm2niix_path, mustWork = F),
  dcm2niix_string
)
}
\arguments{
\item{scanner_type}{MRI scanner vendor type}

\item{input}{Folder path(s) containing dicoms}

\item{output}{Folder path(s) where the nii images should be exported to}

\item{dcm2niix_path}{Path to dcm2niix tool on your system}
}
\value{
List of dcm2niix system commands
}
\description{
Creates the dcm2niix system commands for the conversion
}
\examples{
\dontrun{dcm2nii_wrapper("root_folder/session_id/participant_id/", "nii/session-id/participant-id/")}
}
