% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_setup.R
\name{delete_temp_nii_files}
\alias{delete_temp_nii_files}
\title{Deletes temporary files}
\usage{
delete_temp_nii_files(path_to_search = path_output_converter_temp)
}
\arguments{
\item{path_to_search}{}
}
\description{
Deletes temporary files
}
