% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytable.R
\name{editTable}
\alias{editTable}
\title{Sequence mapper shiny app.}
\usage{
editTable(DF, outdir = getwd(), outfilename = "table")
}
\arguments{
\item{outfilename}{}
}
\value{

}
\description{
Sequence mapper shiny app.
}
