% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_json.R
\name{get_json_headers}
\alias{get_json_headers}
\title{Extracts json headers from multiple files with different headers}
\usage{
get_json_headers(json)
}
\arguments{
\item{json}{a list of json files}

\item{working_dir}{directory}
}
\value{
empty dataframe with each unique column found in one json file
}
\description{
Extracts json headers from multiple files with different headers
}
\examples{
get_json_headers(list_of_jsons)
}
