% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{get_user_input}
\alias{get_user_input}
\title{Start function for the user inputs, asks for input directory, output directory and the order of the input data 'subject_session' or 'session_subject'.}
\usage{
get_user_input()
}
\description{
Start function for the user inputs, asks for input directory, output directory and the order of the input data 'subject_session' or 'session_subject'.
}
