% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dcm2niix.R
\name{install_dcm2niix}
\alias{install_dcm2niix}
\title{installs dcm2niix from Chris Rordens Github Repository, also depending on OS}
\usage{
install_dcm2niix(dcm2niix_release = "v1.0.20211006")
}
\arguments{
\item{dcm2niix_release}{The version number of dcm2niix that you need. E.g. "v1.0.20211006".}
}
\description{
installs dcm2niix from Chris Rordens Github Repository, also depending on OS
}
\examples{
install_dcm2niix()
}
