% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_json.R
\name{list_json_files}
\alias{list_json_files}
\title{This function lists all json files in the folder. It extracts possible sequences based on the "mri_sequences" regex. Furthermore it outputs a "sequence_overview" TSV file containing sequence information and the observations per category. This file is needed for the sequence mapper.}
\usage{
list_json_files(input_path, output_suffix = "")
}
\arguments{
\item{input_path}{File path containing all json files}

\item{output_suffix}{Suffix, which will be added to the "sequence_overview" TSV file.}
}
\description{
This function lists all json files in the folder. It extracts possible sequences based on the "mri_sequences" regex. Furthermore it outputs a "sequence_overview" TSV file containing sequence information and the observations per category. This file is needed for the sequence mapper.
}
