% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_setup.R
\name{path_to_folder}
\alias{path_to_folder}
\title{Create folders from (list of) filenames}
\usage{
path_to_folder(list_of_files)
}
\arguments{
\item{list_of_files}{filename, list of filenames containing path}
}
\value{
Nothing - creates folders for these files on the system
}
\description{
Create folders from (list of) filenames
}
\examples{
path_to_folder("folder/subfolder/file.txt")
}
