% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_settings.R
\name{prepare_environment}
\alias{prepare_environment}
\title{Prepares th environment with custom variables based on the user settings file.}
\usage{
prepare_environment(user_settings_file = settings_file)
}
\arguments{
\item{user_settings_file}{The path to the user settings file, created by "create_user_setting()".}
}
\value{
A mapping of the DICOM to NII mapping
}
\description{
Prepares th environment with custom variables based on the user settings file.
}
