% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{print_passed_time}
\alias{print_passed_time}
\title{Prints progress of files in list of files}
\usage{
print_passed_time(item, list_of_files, start, string)
}
\arguments{
\item{item}{item in for loop}

\item{list_of_files}{list, where the item comes from}

\item{start}{Sys.time() of start, used to calculate the time difference}

\item{string}{String, to describe the function of the loop}
}
\description{
Prints progress of files in list of files
}
