% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_json.R
\name{read_json_headers}
\alias{read_json_headers}
\title{Extracts the information from each json and merges it to the dataframe.
Depends on existing columns! Identified by get_json_headers}
\usage{
read_json_headers(json_path, suffix = "")
}
\arguments{
\item{json_path}{path containing the json files}

\item{output_suffix}{suffix, which will be appended to the output file}
}
\value{
writes a tsv (prefix "json_metadata") containing all headers and the contained information from the JSON files.
}
\description{
Extracts the information from each json and merges it to the dataframe.
Depends on existing columns! Identified by get_json_headers
}
\examples{
read_json_headers(json_list, empty_df_with_headers)
}
