% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input.R
\name{select_user_settings_file}
\alias{select_user_settings_file}
\title{Wrapper function to select and check the user input}
\usage{
select_user_settings_file()
}
\description{
Wrapper function to select and check the user input
}
