% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{start_bids_validator}
\alias{start_bids_validator}
\title{Open BIDS validator in DOCKER}
\usage{
start_bids_validator(bids_path = path_output_bids)
}
\arguments{
\item{bids_path}{The path to the BIDS folder, which is the input file to the BIDS validator}
}
\description{
Open BIDS validator in DOCKER
}
