% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_validation.R
\name{start_bids_validator_docker}
\alias{start_bids_validator_docker}
\title{Open BIDS validator in DOCKER}
\usage{
start_bids_validator_docker(bids_path = user_output_dir)
}
\arguments{
\item{bids_path}{The path to the BIDS folder, which is the input file to the BIDS validator}
}
\value{

}
\description{
Open BIDS validator in DOCKER
}
