% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{write_metadata_bids}
\alias{write_metadata_bids}
\title{This function writes a string to a path in UTF-8.}
\usage{
write_metadata_bids(txt_input, file_path)
}
\arguments{
\item{txt_input}{String input.}

\item{file_path}{Output file path.}
}
\description{
This function writes a string to a path in UTF-8.
}
