% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dicom.R
\name{list_dicom_folders}
\alias{list_dicom_folders}
\title{Finds dicom folders in dicom/session/subject folder structure}
\usage{
list_dicom_folders()
}
\value{
dataframe containing list, session and subject id
}
\description{
Finds dicom folders in dicom/session/subject folder structure
}
\examples{
\dontrun{list_dicom_folders("dicom")}
}
