% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{run_shiny_BIDS}
\alias{run_shiny_BIDS}
\title{Starts the BiDirect BIDS viewer}
\usage{
run_shiny_BIDS(
  shiny_app_path = system.file("rmd", "bidirect_bids_viewer_debug.Rmd", package =
    "BIDSconvertR"),
  bids_directory = path_output_bids
)
}
\arguments{
\item{shiny_app_path}{Path to the viewer RMD file}

\item{bids_directory}{Directory of the images}
}
\value{

}
\description{
Starts the BiDirect BIDS viewer
}
