% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{run_shiny_BIDS}
\alias{run_shiny_BIDS}
\title{Starts the BiDirect BIDS viewer}
\usage{
run_shiny_BIDS(
  bids_directory = shinyDirectoryInput::choose.dir(caption =
    "Please select the BIDS folder."),
  shiny_app_path = system.file("rmd", "bids_viewer.Rmd", package = "bidsconvertr")
)
}
\arguments{
\item{bids_directory}{Directory of the images}

\item{shiny_app_path}{Path to the viewer RMD file}
}
\description{
Starts the BiDirect BIDS viewer
}
